requirements.txt
Django~=5.2.5
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
requests>=2.31,<3.0
pandas>=2.1,<3.0
python-dotenv>=1.0,<2.0
